requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.1
numpy==2.3.2
matplotlib==3.10.5
python-multipart==0.0.20
pydantic==2.11.7
requests==2.32.4
pyyaml==6.0.2
lxml==6.0.0
networkx==3.5
python-dotenv==1.1.1
